Updated ToolService and added CreateLinAgentToolRequest, CreateToolRequest classes

+++ catalogo-strumenti-backend/catalogo-backend-service/src/main/java/it/istat/mec/catalog/request/CreateDocumentationRequest.java
@@ -12,5 +12,5 @@ public class CreateDocumentationRequest {
 	private String documentType;
 	private String notes;	
 	private String resource;
-    private Long tool;
+    private Integer tool;
 }
